Testing moving the fetching of data up in hierarchy, fixed build page not working
